+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.ShoppingCart"
         tools:targetApi="31">
         <activity
-            android:name=".ProductListActivity"
+            android:name=".list.ProductListActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ShoppingCart">
